+++ src/pages/Admin/AdminInstance/AdminInstance.tsx
@@ -5,13 +5,12 @@ import CreateBtn from "@/components/Admin/CreateBtn/CreateBtn";
 import Title from "@/components/Admin/Title/Title";
 import { Pagination } from "antd";
 import { useEffect, useState } from "react";
-import { useLocation, useNavigate } from "react-router-dom";
+import { useLocation } from "react-router-dom";
 import getAdminInstanceListApi from "@/apis/getAdminInstanceListApi";
 import { instanceListDataType, topicDeteilType } from "@/types/adminType";
 import { decrypt } from "@/hooks/useCrypto";
-import postJWTApi from "@/apis/postJWTApi";
-import { PATH } from "@/constants/path";
 import Loading from "@/components/Common/Loading/Loading";
+import { useOnlyAdminPermit } from "@/hooks/queries/useAuthQuery";
 
 const AdminInstance = () => {
   const [instanceModalIsOpen, setInstanceModalIsOpen] =
@@ -22,7 +21,6 @@ const AdminInstance = () => {
   const [totalNumber, setTotalNumber] = useState<number>(0);
   const [isLoading, setIsLoading] = useState(false);
   const location = useLocation();
-  const navigate = useNavigate();
   const handlePageChange = (page: number) => {
     setPageNumber(page);
     setIsLoading(true);
@@ -33,22 +31,13 @@ const AdminInstance = () => {
       setTotalNumber,
     });
   };
+  const { mutate: checkAdmin } = useOnlyAdminPermit();
 
   const topicId = location.state.topicId;
   const decryptTopicId = decrypt(topicId);
 
   useEffect(() => {
-    postJWTApi()
-      .then((res) => {
-        if (res.role !== "ADMIN") {
-          navigate(PATH.ERROR, {
-            state: { errNum: 403, errorTxt: "접근 권한이 없습니다." },
-          });
-        }
-      })
-      .catch(() => {
-        navigate(PATH.LOGIN);
-      });
+    checkAdmin();
     getAdminDetailTopicApi({
       topicId: decryptTopicId,
       setTopicDetail: setTopicDetail,
